requirements.txt
aiofiles==25.1.0
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.34.60
botocore==1.34.162
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
deep-translator==1.11.4
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httptools==0.7.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
limits==5.8.0
mangum==0.21.0
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
pypdf==6.7.5
PyPDF2==3.0.1
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2025.11.3
reportlab==4.4.10
requests==2.32.5
rfc3986==1.5.0
s3transfer==0.10.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
fpdf2
uharfbuzz
fonttools
